scripts/requirements.txt
python-dotenv
requests
inventree